+++ ios/w_client_flutter/Package.swift
@@ -1,0 +1,31 @@
+// swift-tools-version: 5.9
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "w_client_flutter",
+    platforms: [
+        .iOS("12.0")
+    ],
+    products: [
+        .library(name: "w-client-flutter", targets: ["w_client_flutter"])
+    ],
+    dependencies: [
+        .package(name: "FlutterFramework", path: "../FlutterFramework")
+    ],
+    targets: [
+        .target(
+            name: "w_client_flutter",
+            dependencies: [
+                .product(name: "FlutterFramework", package: "FlutterFramework")
+            ],
+            resources: [
+                .process("PrivacyInfo.xcprivacy")
+            ],
+            linkerSettings: [
+                .linkedFramework("UIKit")
+            ]
+        )
+    ]
+)