+++ ios/w_client_flutter/Sources/w_client_flutter/include/WClientSDK.h
@@ -16,4 +16,7 @@
 /// @param parameters 参数字典  需包含orgID、appID、bizSeq、type、miniProgramID（可选）、miniProPgramPlatformID（可选）、uLink（可选）键值对
 - (NSDictionary *)getAuthResult:(NSDictionary *)parameters;
 
+/// 调用网络身份认证App获取认证数据，并在启动失败时回调错误结果
+- (NSDictionary *)getAuthResult:(NSDictionary *)parameters completion:(void (^)(NSDictionary *result))completion;
+
 @end